add new users to Resource Editor group, move signals to separate module
