+++ hms/apps.py
@@ -5,3 +5,6 @@ from django.apps import AppConfig
 
 class HmsConfig(AppConfig):
     name = 'hms'
+
+    def ready(self):
+        from . import signals
